requirements.txt
# requirements.txt
tensorflow-macos>=2.9.0
tensorflow-metal>=0.5.0
flask>=2.0.1
pillow>=8.3.1
numpy>=1.19.5
pandas>=1.3.3
matplotlib>=3.4.3
gtts>=2.2.3
pyttsx3>=2.90
requests>=2.26.0
werkzeug>=2.0.1
